lit_screener/requirements.txt
# Core
streamlit>=1.35.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# LLM providers
openai>=1.30.0
anthropic>=0.28.0

# PDF parsing
PyMuPDF>=1.24.0
pdfplumber>=0.11.0

# Data / Excel
pandas>=2.1.0
openpyxl>=3.1.0

# HTTP (for Drive downloader)
requests>=2.31.0

# Testing
pytest>=8.0.0
